[SPARK-40823][CONNECT] Connect Proto should carry unparsed identifiers
### What changes were proposed in this pull request?

Before this PR, connect proto defines the UnresolvedRelation as
```
  message NamedTable {
    repeated string parts = 1;
  }
```
which asked clients to provide multiple name parts of the relation. For example, user could offer `a.b.c.d` as the table name. However, this actually asks clients to implement `CatalystSqlParser.parseMultipartIdentifier`.  The problem to clients is they cannot access the catalyst parser thus needs to re-invent the wheel. Another problem is clients might not be able to implement the parsing correctly.

This PR proposes to change the proto to
```
  message NamedTable {
    string unparsed_identifier = 1;
}
```
which only needs clients to provide the user's table name. Server side as it can access catalyst, can parse the identifier.

This PR also changes on the Column identifier accordingly.

### Why are the changes needed?

This proposal reduced the required work on the client side.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Existing UT

Closes #38264 from amaliujia/unparsed_identifier.

Authored-by: Rui Wang <rui.wang@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>