[SPARK-40854][CONNECT] Use proper JSON encoding until we have Arrow collection
### What changes were proposed in this pull request?
This patch provides a temporary implementation of result batches as JSON instead of the 'broken' CSV format that was simply generating unescaped CSV lines. In this implementation we actually leverage the existing Spark functionality to generate JSON and then convert this into result batches for Spark Connect.

### Why are the changes needed?
Cleanup

### Does this PR introduce _any_ user-facing change?
No / Experimental

### How was this patch tested?
E2E tests for the Python Client.

Closes #38300 from grundprinzip/spark-json.

Lead-authored-by: Martin Grund <martin.grund@databricks.com>
Co-authored-by: Martin Grund <grundprinzip@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>