[SPARK-40879][CONNECT] Support Join UsingColumns in proto
### What changes were proposed in this pull request?

I was working on refactoring Connect proto tests from Catalyst DSL to DataFrame API, and identified that Join in Connect does not support `UsingColumns`. This is a gap between the Connect proto and DataFrame API. This also blocks the refactoring work because without `UsingColumns`, there is no compatible DataFrame Join API that we can covert existing tests to.

This PR adds the support for Join's `UsingColumns`.

### Why are the changes needed?

1. Improve API coverage.
2. Unblock testing refactoring.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

UT

Closes #38345 from amaliujia/proto-join-using-columns.

Authored-by: Rui Wang <rui.wang@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>