[MINOR][DOC] revisions for spark sql performance tuning to improve readability and grammar
### What changes were proposed in this pull request?

I made some small grammar fixes related to dependent clause followed but independent clauses, starting a sentence with an introductory phrase, using the plural with when are is present in the sentence, and other small fixes to improve readability.

https://spark.apache.org/docs/latest/sql-performance-tuning.html

<img width="1065" alt="Screenshot 2022-11-04 at 15 24 17" src="https://user-images.githubusercontent.com/7563201/199998862-d9418bc1-2fcd-4eff-be8e-af412add6946.png">

### Why are the changes needed?

These changes improve the readability of the Spark documentation for new users or those studying up.

### Does this PR introduce _any_ user-facing change?

Yes, these changes impact the spark documentation.

### How was this patch tested?

No test were created as these changes were solely in markdown.

Closes #38510 from dwsmith1983/minor-doc-revisions.

Lead-authored-by: Dustin William Smith <dustin.smith@deliveryhero.com>
Co-authored-by: dustin <dwsmith1983@users.noreply.github.com>
Co-authored-by: Dustin Smith <Dustin.William.Smith@gmail.com>
Signed-off-by: Sean Owen <srowen@gmail.com>