[SPARK-41001][CONNECT] Make `userId` optional in SparkRemoteSession
### What changes were proposed in this pull request?

This patch provides small changes to the way the `userId` is propagated in the SparkRemoteSession. Previously, the `userId` was a required parameter, this is changed and it's inferred from one of the following configuration options in this order:

1. The `userId` is specified as a parameter to the connection string.
2. The `userId` is specified as a parameter to the `RemoteSparkSession`.
3. If not defined it is inferred from the `$USER`

In addition, since `token` can only be used for secure connections, it will implicitly set the `use_ssl` property of the connection.

### Why are the changes needed?
Usability

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
UT

Closes #38535 from grundprinzip/SPARK-41001_2.

Lead-authored-by: Martin Grund <martin.grund@databricks.com>
Co-authored-by: Martin Grund <grundprinzip@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>