[SPARK-41116][CONNECT] Input relation can be optional for Project in Connect proto
### What changes were proposed in this pull request?

I was writing test cases to test expressions and realized that we can allow `Project` without input plan. For example, `SELECT 1` is a valid query. For SQL it will generate `OneRowRelation` to make up the input plan but for the Connect users they shouldn't need to bother appending that relation. Instead, they can just submit a Project with expressions.

Per our design, Proto is also a API layer and anyone can draft a proto plan without using built-in clients. This PR will improve the proto usability for `Project`.

### Why are the changes needed?

1. Improve usability.
2. Help write test cases for expressions.

### Does this PR introduce _any_ user-facing change?

NO

### How was this patch tested?

UT

Closes #38632 from amaliujia/SPARK-41116.

Authored-by: Rui Wang <rui.wang@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>