[SPARK-41072][SQL][SS] Add the error class `STREAM_FAILED` to `StreamingQueryException`
### What changes were proposed in this pull request?
In the PR, I propose to change the exception `StreamingQueryException` by extending `SparkThrowable` and new error class `STREAM_FAILED` which is set on raising `StreamingQueryException` on any stream failures.

### Why are the changes needed?
To be consistent to other Spark's exception like `SparkException`, and to improve user experience with Spark APIs.

### Does this PR introduce _any_ user-facing change?
Yes, it extends the existing API.

### How was this patch tested?
By running the affected test suites:
```
$ build/sbt "core/testOnly *SparkThrowableSuite"
$ build/sbt "sql/testOnly *QueryExecutionErrorsSuite"
```

Closes #38629 from MaxGekk/StreamingQueryException-SparkThrowable.

Authored-by: Max Gekk <max.gekk@gmail.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>