[SPARK-40657][PROTOBUF] Require shading for Java class jar, improve error handling
### What changes were proposed in this pull request?
The protobuf package shades _com.google.protobuf_ classes. That implies that the user jar providing Java classes also needs to be shaded in the same manner. Otherwise, it does not work as the `Descriptor` class used here and the one used by Java classes in the jar has to match. Otherwise the the Java class can't be instantiated.

This requirement is inconvenient and unintuitive for the users. But we can't make this work without shading as Spark already includes an old version of _protobuf-java_ library (2.5.0) which would conflict with user's jar.

This PR improves error handling to detect various mismatches and provides more descriptive error message.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
- Tested manually with shaded and unshaded jars.
   - How can we test with shaded classes in unit tests?

Closes #38384 from rangadi/shaded-java.

Authored-by: Raghu Angadi <raghu.angadi@databricks.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>