[SPARK-41196][CONNECT][FOLLOW-UP] Fix out of sync generated files for Python
### What changes were proposed in this pull request?

Fix out of sync generated files for Python.

This happens on a rare case for protobuf version change. https://github.com/apache/spark/pull/38693 downgraded protobuf versions.

There were something not generated before but with the protobuf version downgraded that was generated (and this is why there was no merge conflict). However [the downgrading PR](https://github.com/apache/spark/pull/38693) was based on old code before https://github.com/apache/spark/pull/38638 so the protobuf generates based on stale code which leads to stale generated files.

The way to better avoid this is when upon such change, it should lock the repo (partially on some directory to reduce impact), do the work, merge, and enforce pending PR to rebase. However this is not feasible (or too heavy) for concurrent development on Spark repo.

### Why are the changes needed?

Fix out of sync generated files for Python.

### Does this PR introduce _any_ user-facing change?

NO

### How was this patch tested?

UT

Closes #38718 from amaliujia/fix_out_of_sync_proto.

Authored-by: Rui Wang <rui.wang@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>