[SPARK-41300][CONNECT] Unset schema is interpreted as Schema
### What changes were proposed in this pull request?
When a query is read from a DataSource using Spark Connect, the scalar string value would be empty and thus during processing we would treat it as set and fail the query because no schema can be parsed from the empty string.

This patch fixes this issue and adds the relevant test for it.

### Why are the changes needed?
Bugfix

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
UT

Closes #38821 from grundprinzip/SPARK-41300.

Authored-by: Martin Grund <martin.grund@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>