[SPARK-41323][SQL] Support current_schema
### What changes were proposed in this pull request?
Support current_schema(0 as a synonym for current_database()

### Why are the changes needed?
current_schema is used in the SQL Standard and many other products such as Snowflake, Db2, PostgreSQL, Redshift, ...
It also disambiguates from database which is also used with catalog in some products.

### Does this PR introduce _any_ user-facing change?
It is a new feature.

### How was this patch tested?
Add a test verifying that current_schema() and current_database() return the same result()

Closes #38840 from srielau/SPARK-41323-CURRENT_SCHEMA.

Authored-by: Serge Rielau <serge@rielau.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>