[SPARK-41338][SQL] Resolve outer references and normal columns in the same analyzer batch
### What changes were proposed in this pull request?

Today, the way we resolve outer references is very inefficient. It invokes the entire analyzer to resolve the subquery plan, then transforms the plan to resolve `UnresolvedAttribute` to outer references. If the plan is still unresolved, repeat the process until the plan is resolved or the plan doesn't change any more. Ideally, we should only invoke the analyzer once to resolve subquery plans.

This PR adds a new rule to resolve outer references, and put it in the main analyzer batch. Then we can safely invoke the analyzer only once.

### Why are the changes needed?

Simplify the subquery resolution code and make it more efficient

### Does this PR introduce _any_ user-facing change?

no

### How was this patch tested?

existing tests

Closes #38851 from cloud-fan/outer.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>