[SPARK-41437][SQL] Do not optimize the input query twice for v1 write fallback
### What changes were proposed in this pull request?

For v2 DML commands, the input query is a child plan and will be optimized by the optimizer. In the execution path, if the v2 table supports v1 fallback, we will create a DataFrame from the optimized input query plan, and pass the DataFrame to v1 write API.

This means that the input query gets optimized twice: once during the DML command planning, once with the extra DataFrame creation. This is a waste of optimizing time and can cause issues if some custom optimizer rules are not idempotent.

This PR adds a new field `analyzedQuery: Option[LogicalPlan]` to the v2 commands that support v1 fallback. This field will be filled at the end of analysis and will be used to create that v1 fallback DataFrame.

### Why are the changes needed?

avoid double optimizing the input query plan of v2 DML commands

### Does this PR introduce _any_ user-facing change?

no

### How was this patch tested?

new test

Closes #38942 from cloud-fan/v1.

Lead-authored-by: Wenchen Fan <wenchen@databricks.com>
Co-authored-by: Wenchen Fan <cloud0fan@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>